refactor(launch/manager): split render.rs into stage-focused submodules
`render.rs` grew to ~2350 lines and combined the top-level stage
switch, list chrome, editor tabs, modal dispatcher, footer
composition, and visual-regression tests. Split into `render/` as a
directory module:

- `render/mod.rs` — stage-switch entry point (`render`), header
  renderer, shared geometry helpers, cross-cutting re-exports.
- `render/list.rs` — list-stage body, toast, details/sentinel
  right-pane renderers.
- `render/editor.rs` — editor-stage body, tab-bar, per-tab renderers.
- `render/modal.rs` — modal dispatcher (`render_modal`),
  `modal_outer_rect` shared geometry helper.
- `render/footer.rs` — `FooterItem`, `render_footer`.

Tests travel with the behavior they exercise. No visual or layout
change — the operator sees pixel-identical output.

Addresses finding #6 of the PR #166 current-branch review.
